kwtfm.server/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
bcrypt
APScheduler
httpx
httpx-socks
python-jose[cryptography]
redis
pylint
black
requests
debugpy
python-multipart
alembic
python-dateutil